+++ unwired/crypto-common.h
@@ -36,7 +36,7 @@
 
 /*---------------------------------------------------------------------------*/
 
-uint8_t iterator_to_byte(uint8_t iterator);
+uint16_t iterator_to_byte(uint16_t iterator);
 
 void aes_cbc_encrypt(uint32_t *aes_key, uint32_t *nonce, uint32_t *input_data, uint32_t *output_data, uint32_t data_lenth);
 void aes_cbc_decrypt(uint32_t *aes_key, uint32_t *nonce, uint32_t *input_data, uint32_t *output_data, uint32_t data_lenth);
